+++ packages/react-ui/events/src/dom/__tests__/Keyboard-test.internal.js
@@ -158,7 +158,7 @@ describe('Keyboard responder', () => {
     });
 
     // e.g, "Enter" on link
-    test('keyboard click is between key events', () => {
+    test('click is between key events', () => {
       const target = createEventTarget(ref.current);
       target.keydown({key: 'Enter'});
       target.keyup({key: 'Enter'});
@@ -168,7 +168,7 @@ describe('Keyboard responder', () => {
         expect.objectContaining({
           altKey: false,
           ctrlKey: false,
-          defaultPrevented: false,
+          defaultPrevented: true,
           metaKey: false,
           pointerType: 'keyboard',
           shiftKey: false,
@@ -180,7 +180,7 @@ describe('Keyboard responder', () => {
     });
 
     // e.g., "Spacebar" on button
-    test('keyboard click is after key events', () => {
+    test('click is after key events', () => {
       const target = createEventTarget(ref.current);
       target.keydown({key: 'Enter'});
       target.keyup({key: 'Enter'});
@@ -190,7 +190,27 @@ describe('Keyboard responder', () => {
         expect.objectContaining({
           altKey: false,
           ctrlKey: false,
-          defaultPrevented: false,
+          defaultPrevented: true,
+          metaKey: false,
+          pointerType: 'keyboard',
+          shiftKey: false,
+          target: target.node,
+          timeStamp: expect.any(Number),
+          type: 'keyboard:click',
+        }),
+      );
+    });
+
+    // e.g, generated by a screen-reader
+    test('click is orphan', () => {
+      const target = createEventTarget(ref.current);
+      target.virtualclick();
+      expect(onClick).toHaveBeenCalledTimes(1);
+      expect(onClick).toHaveBeenCalledWith(
+        expect.objectContaining({
+          altKey: false,
+          ctrlKey: false,
+          defaultPrevented: true,
           metaKey: false,
           pointerType: 'keyboard',
           shiftKey: false,
@@ -326,6 +346,51 @@ describe('Keyboard responder', () => {
     });
   });
 
+  describe('preventClick', () => {
+    function render(props) {
+      const ref = React.createRef();
+      const Component = () => {
+        const listener = useKeyboard(props);
+        return <div ref={ref} listeners={listener} />;
+      };
+      ReactDOM.render(<Component />, container);
+      return ref;
+    }
+
+    test('prevents native click by default', () => {
+      const onClick = jest.fn();
+      const preventDefault = jest.fn();
+      const ref = render({onClick});
+
+      const target = createEventTarget(ref.current);
+      target.virtualclick({preventDefault});
+
+      expect(preventDefault).toBeCalled();
+      expect(onClick).toHaveBeenCalledTimes(1);
+      expect(onClick).toHaveBeenCalledWith(
+        expect.objectContaining({
+          defaultPrevented: true,
+        }),
+      );
+    });
+
+    test('allows native behaviour if false', () => {
+      const onClick = jest.fn();
+      const preventDefault = jest.fn();
+      const ref = render({onClick, preventClick: false});
+
+      const target = createEventTarget(ref.current);
+      target.virtualclick({preventDefault});
+      expect(preventDefault).not.toBeCalled();
+      expect(onClick).toHaveBeenCalledTimes(1);
+      expect(onClick).toHaveBeenCalledWith(
+        expect.objectContaining({
+          defaultPrevented: false,
+        }),
+      );
+    });
+  });
+
   describe('preventKeys', () => {
     function render(props) {
       const ref = React.createRef();
@@ -347,9 +412,8 @@ describe('Keyboard responder', () => {
       target.keydown({key: 'Tab', preventDefault});
       target.virtualclick({preventDefault: preventDefaultClick});
 
-      expect(onKeyDown).toHaveBeenCalledTimes(1);
       expect(preventDefault).toBeCalled();
-      expect(preventDefaultClick).toBeCalled();
+      expect(onKeyDown).toHaveBeenCalledTimes(1);
       expect(onKeyDown).toHaveBeenCalledWith(
         expect.objectContaining({
           defaultPrevented: true,
@@ -362,19 +426,12 @@ describe('Keyboard responder', () => {
     test('key config matches (modifier keys)', () => {
       const onKeyDown = jest.fn();
       const preventDefault = jest.fn();
-      const preventDefaultClick = jest.fn();
       const ref = render({onKeyDown, preventKeys: [['Tab', {shiftKey: true}]]});
 
       const target = createEventTarget(ref.current);
       target.keydown({key: 'Tab', preventDefault, shiftKey: true});
-      target.virtualclick({
-        preventDefault: preventDefaultClick,
-        shiftKey: true,
-      });
-
-      expect(onKeyDown).toHaveBeenCalledTimes(1);
       expect(preventDefault).toBeCalled();
-      expect(preventDefaultClick).toBeCalled();
+      expect(onKeyDown).toHaveBeenCalledTimes(1);
       expect(onKeyDown).toHaveBeenCalledWith(
         expect.objectContaining({
           defaultPrevented: true,
@@ -388,19 +445,12 @@ describe('Keyboard responder', () => {
     test('key config does not match (modifier keys)', () => {
       const onKeyDown = jest.fn();
       const preventDefault = jest.fn();
-      const preventDefaultClick = jest.fn();
       const ref = render({onKeyDown, preventKeys: [['Tab', {shiftKey: true}]]});
 
       const target = createEventTarget(ref.current);
       target.keydown({key: 'Tab', preventDefault, shiftKey: false});
-      target.virtualclick({
-        preventDefault: preventDefaultClick,
-        shiftKey: false,
-      });
-
-      expect(onKeyDown).toHaveBeenCalledTimes(1);
       expect(preventDefault).not.toBeCalled();
-      expect(preventDefaultClick).not.toBeCalled();
+      expect(onKeyDown).toHaveBeenCalledTimes(1);
       expect(onKeyDown).toHaveBeenCalledWith(
         expect.objectContaining({
           defaultPrevented: false,
