+++ packages/react-ui/npm/drag.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-events-drag.production.min.js');
+  module.exports = require('./cjs/react-ui-events/drag.production.min.js');
 } else {
-  module.exports = require('./cjs/react-events-drag.development.js');
+  module.exports = require('./cjs/react-ui-events/drag.development.js');
 }
