+++ packages/react-ui/npm/focus.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-events-focus.production.min.js');
+  module.exports = require('./cjs/react-ui-events/focus.production.min.js');
 } else {
-  module.exports = require('./cjs/react-events-focus.development.js');
+  module.exports = require('./cjs/react-ui-events/focus.development.js');
 }
