+++ packages/react-ui/npm/hover.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-events-hover.production.min.js');
+  module.exports = require('./cjs/react-ui-events/hover.production.min.js');
 } else {
-  module.exports = require('./cjs/react-events-hover.development.js');
+  module.exports = require('./cjs/react-ui-events/hover.development.js');
 }
