+++ packages/react-ui/npm/input.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-events-input.production.min.js');
+  module.exports = require('./cjs/react-ui-events/input.production.min.js');
 } else {
-  module.exports = require('./cjs/react-events-input.development.js');
+  module.exports = require('./cjs/react-ui-events/input.development.js');
 }
