+++ packages/react-ui/npm/keyboard.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-events-keyboard.production.min.js');
+  module.exports = require('./cjs/react-ui-events/keyboard.production.min.js');
 } else {
-  module.exports = require('./cjs/react-events-keyboard.development.js');
+  module.exports = require('./cjs/react-ui-events/keyboard.development.js');
 }
