+++ packages/react-ui/npm/press.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-events-press.production.min.js');
+  module.exports = require('./cjs/react-ui-events/press.production.min.js');
 } else {
-  module.exports = require('./cjs/react-events-press.development.js');
+  module.exports = require('./cjs/react-ui-events/press.development.js');
 }
