+++ packages/react-ui/npm/scroll.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-events-scroll.production.min.js');
+  module.exports = require('./cjs/react-ui-events/scroll.production.min.js');
 } else {
-  module.exports = require('./cjs/react-events-scroll.development.js');
+  module.exports = require('./cjs/react-ui-events/scroll.development.js');
 }
