+++ packages/react-ui/npm/swipe.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-events-swipe.production.min.js');
+  module.exports = require('./cjs/react-ui-events/swipe.production.min.js');
 } else {
-  module.exports = require('./cjs/react-events-swipe.development.js');
+  module.exports = require('./cjs/react-ui-events/swipe.development.js');
 }
