+++ packages/react-ui/npm/tap.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-events-tap.production.min.js');
+  module.exports = require('./cjs/react-ui-events/tap.production.min.js');
 } else {
-  module.exports = require('./cjs/react-events-tap.development.js');
+  module.exports = require('./cjs/react-ui-events/tap.development.js');
 }
